All updated raghu

+++ backend/src/main/java/com/medi360/service/NotificationService.java
@@ -10,6 +10,8 @@ import org.springframework.stereotype.Service;
 import com.medi360.db.NotificationRepository;
 import com.medi360.entities.Notification;
 import com.medi360.entities.User;
+import com.medi360.exception.NotificationNotfoundException;
+import com.medi360.exception.UserNotFoundException;
 
 @Service
 public class NotificationService {
@@ -21,11 +23,19 @@ public class NotificationService {
 		return this.notificationrepo.save(notification);
 	}
 	
-	public Notification updateNotification(Notification notification) {
+	public Notification updateNotification(Notification notification) throws NotificationNotfoundException{
+		if (!notificationrepo.existsById(notification.getNotificationId())) {
+			throw new NotificationNotfoundException("Notification not found with id " + notification.getNotificationId());
+		}
 		return this.notificationrepo.save(notification);
 	}
 	
-	public String deleteNotification(int id) {
+	public String deleteNotification(int id) throws NotificationNotfoundException{
+		
+		if (!notificationrepo.existsById(id)) {
+			throw new NotificationNotfoundException("Notification not found with id " + id);
+		}
+		
 		this.notificationrepo.deleteById(id);
 		return "Successfully deleted";
 	}
@@ -37,4 +47,12 @@ public class NotificationService {
 	public Page<Notification> getAllNotificationsWithPagination(Pageable pageable) {
 		return this.notificationrepo.findAll(pageable);
 	}
+	
+	public Notification findById(int id) throws NotificationNotfoundException{
+		if (!notificationrepo.existsById(id)) {
+			throw new NotificationNotfoundException("Notification not found with id " + id);
+		}
+		return this.notificationrepo.findById(id).get();
+	}
+	
 }
